split query package
