+++ core/query_test.go
@@ -66,323 +66,6 @@ func TestQueryErrorStringExtraMessage(t *testing.T) {
 	}
 }
 
-func TestUnmarshalResponseMapToStruct(t *testing.T) {
-	// given
-	m := []map[string]interface{}{{
-		"Tag1": "test1",
-		"Tag2": "test2",
-	}}
-	have := struct {
-		Tag1 string
-		Tag2 string
-	}{}
-
-	want := struct {
-		Tag1 string
-		Tag2 string
-	}{
-		Tag1: "test1",
-		Tag2: "test2",
-	}
-
-	// when
-	err := core.UnmarshalResponse(m, &have)
-
-	// then
-	if err != nil {
-		LogTestError(have, want, t, fmt.Sprintf("Test failed with error: %+v", err))
-	}
-	if have != want {
-		LogTestError(have, want, t)
-	}
-}
-
-func TestUnmarshalResponseMapToStructError(t *testing.T) {
-	// given
-	m := []map[string]interface{}{{
-		"Tag1": "test1",
-		"Tag2": "test2",
-	}}
-	have := struct {
-		Tag1 []bool
-		Tag2 string
-	}{}
-	want := "1 error(s) decoding:\n\n* cannot parse 'Tag1[0]' as bool: strconv.ParseBool: parsing \"test1\": invalid syntax"
-
-	// when
-	err := core.UnmarshalResponse(m, &have)
-
-	// then
-	if err == nil {
-		LogTestError(have, want, t, "expected unmarshal error for type []bool")
-	}
-	if err.Error() != want {
-		LogTestError(err.Error(), want, t)
-	}
-}
-
-func TestUnmarshalResponseMapToArray(t *testing.T) {
-	// given
-	m := []map[string]interface{}{{
-		"Tag1": "test1_1",
-		"Tag2": "test1_2",
-	}, {
-		"Tag1": "test2_1",
-		"Tag2": "test2_2",
-	}}
-	have := [2]struct {
-		Tag1 string
-		Tag2 string
-	}{}
-	want := [2]struct {
-		Tag1 string
-		Tag2 string
-	}{{
-		Tag1: "test1_1",
-		Tag2: "test1_2",
-	}, {
-		Tag1: "test2_1",
-		Tag2: "test2_2",
-	}}
-
-	// when
-	err := core.UnmarshalResponse(m, &have)
-
-	// then
-	if err != nil {
-		LogTestError(have, want, t, fmt.Sprintf("test failed with error %s", err))
-	}
-	for i, v := range have {
-		if v.Tag1 != want[i].Tag1 {
-			LogTestError(v.Tag1, want[i].Tag1, t)
-		}
-		if v.Tag2 != want[i].Tag2 {
-			LogTestError(v.Tag2, want[i].Tag2, t)
-		}
-	}
-}
-
-func TestUnmarshalResponseMapToShorterArray(t *testing.T) {
-	// given
-	m := []map[string]interface{}{{
-		"Tag1": "test1_1",
-		"Tag2": "test1_2",
-	}, {
-		"Tag1": "test2_1",
-		"Tag2": "test2_2",
-	}}
-	have := [1]struct {
-		Tag1 string
-		Tag2 string
-	}{}
-	want := [1]struct {
-		Tag1 string
-		Tag2 string
-	}{{
-		Tag1: "test1_1",
-		Tag2: "test1_2",
-	}}
-
-	// when
-	err := core.UnmarshalResponse(m, &have)
-
-	// then
-	if err != nil {
-		LogTestError(have, want, t, fmt.Sprintf("test failed with error %s", err))
-	}
-	for i, v := range have {
-		if v.Tag1 != want[i].Tag1 {
-			LogTestError(v.Tag1, want[i].Tag1, t)
-		}
-		if v.Tag2 != want[i].Tag2 {
-			LogTestError(v.Tag2, want[i].Tag2, t)
-		}
-	}
-}
-
-func TestUnmarshalResponseMapToArrayError(t *testing.T) {
-	// given
-	m := []map[string]interface{}{{
-		"Tag1": "test1_1",
-		"Tag2": "test1_2",
-	}, {
-		"Tag1": "test2_1",
-		"Tag2": "test2_2",
-	}}
-	have := [2]struct {
-		Tag1 []bool
-		Tag2 string
-	}{}
-	want := "1 error(s) decoding:\n\n* cannot parse 'Tag1[0]' as bool: strconv.ParseBool: parsing \"test1_1\": invalid syntax"
-
-	// when
-	err := core.UnmarshalResponse(m, &have)
-
-	// then
-	if err == nil {
-		LogTestError(have, want, t, "expected error unmarshaling")
-	}
-	if err.Error() != want {
-		LogTestError(err.Error(), want, t)
-	}
-}
-
-func TestUnmarshalResponseMapToSlice(t *testing.T) {
-	// given
-	m := []map[string]interface{}{{
-		"Tag1": "test1_1",
-		"Tag2": "test1_2",
-	}, {
-		"Tag1": "test2_1",
-		"Tag2": "test2_2",
-	}}
-	have := []struct {
-		Tag1 string
-		Tag2 string
-	}{}
-	want := []struct {
-		Tag1 string
-		Tag2 string
-	}{{
-		Tag1: "test1_1",
-		Tag2: "test1_2",
-	}, {
-		Tag1: "test2_1",
-		Tag2: "test2_2",
-	}}
-
-	// when
-	err := core.UnmarshalResponse(m, &have)
-
-	// then
-	if err != nil {
-		LogTestError(have, want, t, fmt.Sprintf("test failed with error %s", err))
-	}
-	for i, v := range have {
-		if v.Tag1 != want[i].Tag1 {
-			LogTestError(v.Tag1, want[i].Tag1, t)
-		}
-		if v.Tag2 != want[i].Tag2 {
-			LogTestError(v.Tag2, want[i].Tag2, t)
-		}
-	}
-}
-
-func TestUnmarshalResponseMapToSliceWithContent(t *testing.T) {
-	// given
-	m := []map[string]interface{}{{
-		"Tag1": "test1_1",
-		"Tag2": "test1_2",
-	}, {
-		"Tag1": "test2_1",
-		"Tag2": "test2_2",
-	}}
-	have := []struct {
-		Tag1 string
-		Tag2 string
-	}{{
-		Tag1: "do not",
-		Tag2: "override",
-	}}
-	want := []struct {
-		Tag1 string
-		Tag2 string
-	}{{
-		Tag1: "do not",
-		Tag2: "override",
-	}, {
-		Tag1: "test1_1",
-		Tag2: "test1_2",
-	}, {
-		Tag1: "test2_1",
-		Tag2: "test2_2",
-	}}
-
-	// when
-	err := core.UnmarshalResponse(m, &have)
-
-	// then
-	if err != nil {
-		LogTestError(have, want, t, fmt.Sprintf("test failed with error %s", err))
-	}
-	for i, v := range have {
-		if v.Tag1 != want[i].Tag1 {
-			LogTestError(v.Tag1, want[i].Tag1, t)
-		}
-		if v.Tag2 != want[i].Tag2 {
-			LogTestError(v.Tag2, want[i].Tag2, t)
-		}
-	}
-}
-
-func TestUnmarshalResponseMapToSliceError(t *testing.T) {
-	// given
-	m := []map[string]interface{}{{
-		"Tag1": "test1_1",
-		"Tag2": "test1_2",
-	}, {
-		"Tag1": "test2_1",
-		"Tag2": "test2_2",
-	}}
-	have := []struct {
-		Tag1 []bool
-		Tag2 string
-	}{}
-	want := "1 error(s) decoding:\n\n* cannot parse 'Tag1[0]' as bool: strconv.ParseBool: parsing \"test1_1\": invalid syntax"
-
-	// when
-	err := core.UnmarshalResponse(m, &have)
-
-	// then
-	if err == nil {
-		LogTestError(have, want, t, "expected error unmarshaling")
-	}
-	if err.Error() != want {
-		LogTestError(have, want, t)
-	}
-}
-
-func TestUnmarshalResponseMapToUnsupportedTypeError(t *testing.T) {
-	// given
-	m := []map[string]interface{}{{
-		"Tag1": "test",
-	}}
-	have := ""
-
-	// when
-	err := core.UnmarshalResponse(m, &have)
-
-	// then
-	if err == nil {
-		LogTestError(have, "", t, "expected unmarshal error")
-	}
-	if err.Error() != "unsupported type string. Expected type struct, slice or array" {
-		LogTestError(err.Error(), "unsupported type string. Expected type struct, slice or array", t)
-	}
-}
-
-func TestDecodeMapToNonPtrError(t *testing.T) {
-	// given
-	m := map[string]interface{}{
-		"Tag1": "test",
-	}
-	have := struct {
-		Tag1 string
-	}{}
-	want := "expected pointer to value, not value"
-
-	// when
-	err := core.Decode(m, have)
-
-	// then
-	if err == nil {
-		LogTestError(have, want, t, "expected error because of non pointer value")
-	}
-	if err.Error() != want {
-		LogTestError(err.Error(), want, t)
-	}
-}
-
 func TestKeepAlive(t *testing.T) {
 	// given
 	b := &bytes.Buffer{}
@@ -477,8 +160,8 @@ func TestRunSimpleResponse(t *testing.T) {
 		LogTestError(data, "", t, "didn't expect return value")
 	}
 	have := writer.String()
-	if have != string(cmd) {
-		LogTestError(have, cmd, t)
+	if have != fmt.Sprintf("%s\n", cmd) {
+		LogTestError(have, fmt.Sprintf("%s\n", cmd), t)
 	}
 }
 
@@ -593,14 +276,14 @@ func TestRunUnexpectedError(t *testing.T) {
 	}
 }
 
-func TestDemultiplexer(t *testing.T) {
+func TestSplit(t *testing.T) {
 	// given
 	out := make(chan []byte)
 	notify := make(chan []byte)
 	in := bytes.NewBufferString("\rtest\n\rerror id=0 msg=\n\rnotify test\n\r")
 
 	// when
-	go core.Demultiplexer(in, out, notify)
+	go core.Split(in, out, notify)
 
 	// then
 	go func() {
